+++ R/diagnostic_posterior.R
@@ -3,9 +3,12 @@
 #' Extract diagnostic metrics (Effective Sample Size (`ESS`), `Rhat` and Monte
 #' Carlo Standard Error `MCSE`).
 #'
-#' @param posterior A `stanreg`, `stanfit`, `brmsfit`, or `blavaan` object.
+#' @param posterior A `stanreg`, `stanfit`, `brmsfit`, or `blavaan` object; a
+#' list of data frames or matrices representing MCMC chains (rows as samples,
+#' columns as parameters); or a 3D array (dimensions: samples, chains,
+#' parameters)
 #' @param diagnostic Diagnostic metrics to compute.  Character (vector) or list
-#'   with one or more of these options: `"ESS"`, `"Rhat"`, `"MCSE"` or `"all"`.
+#' with one or more of these options: `"ESS"`, `"Rhat"`, `"MCSE"` or `"all"`.
 #'
 #' @inheritSection hdi Model components
 #'
@@ -42,6 +45,15 @@
 #' model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
 #' diagnostic_posterior(model)
 #' }
+#' @examplesIf require("rstan")
+#' set.seed(101)
+#' mkdata <- function(nrow = 1000, ncol = 2, parnm = LETTERS[1:ncol]) {
+#'   x <- as.data.frame(replicate(ncol, rnorm(nrow)))
+#'   names(x) <- parnm
+#'   x
+#' }
+#' dd <- replicate(5, mkdata(), simplify = FALSE)
+#' diagnostic_posterior(dd)
 #' @references
 #' - Gelman, A., & Rubin, D. B. (1992). Inference from iterative simulation
 #'   using multiple sequences. Statistical science, 7(4), 457-472.
@@ -58,20 +70,62 @@ diagnostic_posterior <- function(posterior, ...) {
 
 #' @rdname diagnostic_posterior
 #' @export
-diagnostic_posterior.default <- function(posterior, diagnostic = c("ESS", "Rhat"), ...) {
-  insight::format_error("'diagnostic_posterior()' only works with rstanarm, brms or blavaan models.")
+diagnostic_posterior.default <- function(posterior, diagnostic = "all", ...) {
+  ## check input, coerce to array
+  if (is.list(posterior)) {
+    for (i in seq_along(posterior)) {
+      p <- posterior[[i]]
+      if (
+        !((inherits(p, "data.frame") || inherits(p, "matrix")) &&
+          (length(dim(p)) == 2) &&
+          (ncol(p) == ncol(posterior[[1]])))
+      ) {
+        insight::format_error(
+          "'posterior' must be a 3D array or a list of data frames with equal numbers of columns."
+        )
+      }
+    }
+    insight::check_if_installed("posterior")
+    posterior <- posterior::as_draws_array(posterior)
+    ## draws_array() class messes things up downstream ...
+    class(posterior) <- "array"
+  }
+  if (!(inherits(posterior, "array") && length(dim(posterior)) == 3)) {
+    insight::format_error("Expecting a 3D array for 'posterior'.")
+  }
+
+  ret <- data.frame(Parameter = colnames(posterior[[1]]))
+  if (is.null(diagnostic)) {
+    return(ret)
+  }
+
+  .diag_opts <- c("Rhat", "ESS", "MCSE")
+  if (diagnostic == "all") {
+    diagnostic <- .diag_opts
+  }
+
+  ## need ESS for MCSE, so compute these in any case
+  insight::check_if_installed("rstan")
+  mon <- rstan::monitor(posterior, print = FALSE, probs = 0.5)
+  ret <- with(
+    mon,
+    data.frame(Parameter = rownames(mon), ESS = n_eff, Rhat, MCSE = MCSE_Q50)
+  )
+  ret[c("Parameter", diagnostic)]
 }
 
 #' @inheritParams insight::get_parameters.BFBayesFactor
 #' @inheritParams insight::get_parameters
 #' @rdname diagnostic_posterior
 #' @export
-diagnostic_posterior.stanreg <- function(posterior,
-                                         diagnostic = "all",
-                                         effects = "fixed",
-                                         component = "location",
-                                         parameters = NULL,
-                                         ...) {
+diagnostic_posterior.stanreg <- function(
+  posterior,
+  diagnostic = "all",
+  effects = "fixed",
+  component = "location",
+  parameters = NULL,
+  ...
+) {
   # Find parameters
   params <- insight::find_parameters(
     posterior,
@@ -125,11 +179,13 @@ diagnostic_posterior.stanreg <- function(posterior,
 
 #' @inheritParams insight::get_parameters
 #' @export
-diagnostic_posterior.stanmvreg <- function(posterior,
-                                           diagnostic = "all",
-                                           effects = "fixed",
-                                           parameters = NULL,
-                                           ...) {
+diagnostic_posterior.stanmvreg <- function(
+  posterior,
+  diagnostic = "all",
+  effects = "fixed",
+  parameters = NULL,
+  ...
+) {
   # Find parameters
   all_params <- insight::find_parameters(
     posterior,
@@ -138,10 +194,13 @@ diagnostic_posterior.stanmvreg <- function(posterior,
     flatten = FALSE
   )
 
-  params <- unlist(lapply(names(all_params), function(i) {
-    all_params[[i]]$sigma <- NULL
-    unlist(all_params[[i]], use.names = FALSE)
-  }), use.names = FALSE)
+  params <- unlist(
+    lapply(names(all_params), function(i) {
+      all_params[[i]]$sigma <- NULL
+      unlist(all_params[[i]], use.names = FALSE)
+    }),
+    use.names = FALSE
+  )
 
   # If no diagnostic
   if (is.null(diagnostic)) {
@@ -201,14 +260,17 @@ diagnostic_posterior.stanmvreg <- function(posterior,
 
 #' @inheritParams insight::get_parameters
 #' @export
-diagnostic_posterior.brmsfit <- function(posterior,
-                                         diagnostic = "all",
-                                         effects = "fixed",
-                                         component = "conditional",
-                                         parameters = NULL,
-                                         ...) {
+diagnostic_posterior.brmsfit <- function(
+  posterior,
+  diagnostic = "all",
+  effects = "fixed",
+  component = "conditional",
+  parameters = NULL,
+  ...
+) {
   # Find parameters
-  params <- insight::find_parameters(posterior,
+  params <- insight::find_parameters(
+    posterior,
     effects = effects,
     component = component,
     parameters = parameters,
@@ -259,7 +321,13 @@ diagnostic_posterior.brmsfit <- function(posterior,
 
 #' @inheritParams insight::get_parameters
 #' @export
-diagnostic_posterior.stanfit <- function(posterior, diagnostic = "all", effects = "fixed", parameters = NULL, ...) {
+diagnostic_posterior.stanfit <- function(
+  posterior,
+  diagnostic = "all",
+  effects = "fixed",
+  parameters = NULL,
+  ...
+) {
   # Find parameters
   params <- insight::find_parameters(
     posterior,
@@ -285,10 +353,7 @@ diagnostic_posterior.stanfit <- function(posterior, diagnostic = "all", effects 
 
   insight::check_if_installed("rstan")
 
-  all_params <- insight::find_parameters(posterior,
-    effects = effects,
-    flatten = TRUE
-  )
+  all_params <- insight::find_parameters(posterior, effects = effects, flatten = TRUE)
 
   diagnostic_df <- data.frame(
     Parameter = all_params,
@@ -356,7 +421,6 @@ diagnostic_posterior.blavaan <- function(posterior, diagnostic = "all", ...) {
     ESS <- effective_sample(posterior)
     out <- merge(out, ESS, by = "Parameter", all = TRUE)
   }
-
 
   if ("MCSE" %in% diagnostic) {
     MCSE <- mcse(posterior)